refactor(vendo): sync is the shared flow's incremental wrapper
The flow-level cases (AI-flag matrix, theme re-scan, pin baselines) move to
sync-flow.test.ts retargeted at runSyncFlow; sync.coherence.test.ts keeps only
what the WRAPPER owns (who counts as interactive, one-object stdout).
